Use custom fonts
Font-awesome although is in fact awesome, is massive. We're not using
every single icon, in fact, we're using only 4 out of 675... It's
unreasonable to keep using it.

We're preparing our own font and styling, to keep the same approach,
with much less MB to be loaded.

It also, fixes the issue with [go-libsas][1] we're having, allowing us
to move forward with the application.

[1]: https://github.com/wellington/go-libsass/issues/52