Apply fixes
The menu has been modified on mobile. It does no longer require JS nor a
listed style view. We can safely display both links next to each other.

Bourbon is changing. Not sure if I like it or not. Some of the
functionality is being depricated. Tackling the REM converter in this
commit should prevent us from getting most of the warnings while
compiling.