Modernise styles
The world is moving away from things like bourbon, which initially
brought some functionality to make the styles compatible with different
web browsers. The cool kids nowadays use PostCSS technology to deal with
that. I'm replacing the bourbon methods with native CSS properties and
enabling the css-autoprefix to fix the styles for anyone using a web
browser from this decade.