Make improvements
Adding some more info in regards to the timing working at certain
places. I've learned this may be important to some people, therefore
adaptation in order.

Also, workplaces now have a list of duties/achievements assigned to
them, so if someone is interested, they don't have to reach out if they
don't want to.

Fixing some styling issues for Firefox. Unfortunately the progress boxes
look different. Adding extra styling parameters brings them together to
some extent. So it happens, the detail section is different too.
Removing it once again for Firefox.

Upgrading the scales and reordering some of the "other" technologies.
Times change, meaning different techs are more often used than the
other.

Before each build & run, download an avatar and store it in our files.
This should help with performance whilst no longer loading from external
source and being able to set own headers for that image.

Speaking of performance, we're also getting rid of the Lato font in
favour of some default ones, such as Nato or Helvetica.

The default `details` behaviour is not as flexible as one would hope
for. We're re-implementing it with the use of JS and CSS keeping the
HTML syntax intact. This also means IE/Edge support.